+++ changelog.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## Upcoming
-* Added "Pumped" effect, unused as now
+* Added "Pumped" effect, given when the player sleeps and doesn't have the Tired effect
   * Decreases hunger consumption
 * Added `Arrows don't trigger invincibility frames`.
   * Like combat test snapshots, hitting a mob with multiple arrows will no longer bounce them off but actually hit it.